feat(notification): use composables for notification functionality in router views
